[REF] Move functions and classes into appropriate modules (#489)
* Move get_atlas_[n|c]ifti into utils.atlas.

* Move zscore_nifti into utils.utils.

* Move butter_bandpass into utils.utils.

* Move helper functions into utils.filemanip.

* Move regression funcs into utils.utils.

* Move CiftiCorrelation from utils to interfaces.

* Move CiftiParcellate from utils to interfaces.

* Move CiftiSurfaceResample from utils to interfaces.

* Move CiftiSeparateMetric from utils to interfaces.

* Run isort.

* Fix input/output specs.

* Update connectivity.py

* Move utils.restingstate to interfaces.resting_state.

* Fix imports.

* Move DerivativesDataSink to interfaces.
