+++ xcp_d/workflow/anatomical.py
@@ -22,6 +22,7 @@ from niworkflows.engine.workflows import LiterateWorkflow as Workflow
 from templateflow.api import get as get_template
 
 from xcp_d.interfaces.ants import CompositeInvTransformUtil, ConvertTransformFile
+from xcp_d.interfaces.bids import DerivativesDataSink as BIDSDerivativesDataSink
 from xcp_d.interfaces.c3 import C3d  # TM
 from xcp_d.interfaces.connectivity import ApplyTransformsx
 from xcp_d.interfaces.surfplotting import BrainPlotx, RibbontoStatmap
@@ -29,14 +30,13 @@ from xcp_d.interfaces.workbench import (  # MB,TM
     ApplyAffine,
     ApplyWarpfield,
     ChangeXfmType,
+    CiftiSurfaceResample,
     ConvertAffine,
     SurfaceAverage,
     SurfaceGenerateInflated,
     SurfaceSphereProjectUnproject,
 )
-from xcp_d.utils.bids import DerivativesDataSink as BIDSDerivativesDataSink
 from xcp_d.utils.bids import collect_data
-from xcp_d.utils.ciftiresample import CiftiSurfaceResample
 
 
 class DerivativesDataSink(BIDSDerivativesDataSink):
